requirements-cli.txt
click>=8.0.0
python-dotenv>=0.19.0
requests>=2.25.0
